SIMD with AVX2 on supported AMD64 machines
Some help from gpt-5.1-thinking taken: used wrong register size for
the weighted sum at first, so it was truncating the second half of
our block; also there was an overflow from the modulus and stuff.

Unfortunately the AVX2 adler32 is only about 20% faster than the
generic version which doesn't make for much.